pyproject.toml
[tool.poetry]
name = "ndai"
version = "0.1.0"
description = ""
authors = ["Ben Maheu <benjamin.maheu@student.ecp.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.13"
langchain-text-splitters = "^0.3.11"
pytesseract = "^0.3.13"
pillow = "^12.0.0"
pdf2image = "^1.17.0"
pymupdf = "^1.26.5"
flask = "^3.1.2"
openai = "^2.4.0"
dotenv = "^0.9.9"
chromadb = "^1.1.1"
torch = { version = "2.2.2", source="torch"}
sentence_transformers = "^5.1.1"
numpy = "<2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu116"
secondary = true